chore: move lambda folder
