+++ src/cores/lambda/lambda-cognito-handler.ts
@@ -9,7 +9,7 @@
  * @copyright (C) 2019 LemonCloud Co Ltd. - All Rights Reserved.
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { $engine, _log, _inf, _err, $U, $_ } from '../engine/';
+import { $engine, _log, _inf, _err, $U, $_ } from '../../engine/';
 const NS = $U.NS('HCOG', 'yellow'); // NAMESPACE TO BE PRINTED.
 
 import $lambda, { LambdaHandler, CognitoHandler, LambdaHandlerService } from './lambda-handler';
