+++ src/cores/lambda/lambda-dynamo-stream-handler.spec.ts
@@ -8,14 +8,14 @@
  *
  * @copyright (C) 2019 LemonCloud Co Ltd. - All Rights Reserved.
  */
-import { expect2, GETERR } from '../common/test-helper';
-import { loadJsonSync } from '../tools/';
+import { expect2, GETERR } from '../../common/test-helper';
+import { loadJsonSync } from '../../tools/';
 
 import * as $lambda from './lambda-handler.spec';
-import * as $elastic6 from './elastic6-service.spec';
+import * as $elastic6 from './../elastic6-service.spec';
 import { LambdaDynamoStreamHandler } from './lambda-dynamo-stream-handler';
 import { LambdaHandler } from './lambda-handler';
-import { DynamoOption } from './dynamo-service';
+import { DynamoOption } from './../dynamo-service';
 
 class LambdaDynamoStreamHandlerLocal extends LambdaDynamoStreamHandler {
     public constructor(lambda: LambdaHandler) {
