+++ src/cores/lambda/lambda-web-handler.spec.ts
@@ -8,12 +8,12 @@
  *
  * @copyright (C) 2019 LemonCloud Co Ltd. - All Rights Reserved.
  */
-import { expect2, GETERR$ } from '../common/test-helper';
-import { $U } from '../engine/';
-import { loadJsonSync } from '../tools/';
+import { expect2, GETERR$ } from '../../common/test-helper';
+import { $U } from '../../engine/';
+import { loadJsonSync } from '../../tools/';
 
 import * as $lambda from './lambda-handler.spec';
-import { NextDecoder, NextHandler, NextContext, ProtocolParam } from './core-types';
+import { NextDecoder, NextHandler, NextContext, ProtocolParam } from './../core-types';
 import { LambdaWEBHandler } from './lambda-web-handler';
 import { LambdaHandler } from './lambda-handler';
